:sparkles: #130 Store unlocked installation in store with slug
- Store unlocked installation in store
- When user navigate to installation details, we check if installation slug is stored in the store and then display the correct component
