+++ client/src/pages/Installation/installation.scss
@@ -4,6 +4,9 @@
   background-color: $color-blue;
   height: 100vh;
 
+  color: white;
+  font-size: 16px;
+
   &__wrapper {
     height: 100%;
     width: 100%;
